convert to TSX, construct out
