+++ packages/commonwealth/client/scripts/views/pages/new_snapshot_proposal/index.tsx
@@ -6,7 +6,7 @@ import ClassComponent from 'class_component';
 import 'pages/new_proposal_page.scss';
 
 import Sublayout from 'views/sublayout';
-import NewProposalForm from './new_proposal_form';
+import { NewProposalForm } from './new_proposal_form';
 
 type NewProposalSnapshotPageAttrs = { snapshotId: string };
 
@@ -17,7 +17,7 @@ class NewSnapshotProposalPage extends ClassComponent<NewProposalSnapshotPageAttr
       // title="New Snapshot Proposal"
       >
         <div class="NewProposalPage">
-          {m(NewProposalForm, { snapshotId: vnode.attrs.snapshotId })}
+          <NewProposalForm snapshotId={vnode.attrs.snapshotId} />
         </div>
       </Sublayout>
     );
